requirements.txt
flask==0.12.2
Pillow==7.2.0
matplotlib==3.2.2
numpy==1.18.5
opencv-python
